fixed int vs float bug

+++ helpers.py
@@ -159,10 +159,10 @@ class Volpkg(object):
 
 		H = len(interpolated)
 		W = max([len(i) for i in interpolated])
-		im = np.zeros((H, W*2, 3), dtype=np.uint8)
-		#replace all 0's with [nan, nan, nan]
+		im = np.zeros((H, W*2, 3), dtype=np.float64)
+		#replace all 0's with nan
 		im[im == 0] = np.nan
-		
+
 		cy = H//2
 
 		Center = interpolated[cy][len(interpolated[cy])//2]
